Use decimal instead of float, to get consistent rounding (#553)

+++ custom_components/powercalc/sensors/group.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+from decimal import Decimal
 
 from homeassistant.components.sensor import DOMAIN as SENSOR_DOMAIN
 from homeassistant.components.sensor import (
@@ -58,7 +59,7 @@ class GroupedSensor(SensorEntity):
         states: list[State] = list(filter(None, all_states))
         ignored_states = (STATE_UNAVAILABLE, STATE_UNKNOWN)
         summed = sum(
-            float(state.state) for state in states if state.state not in ignored_states
+            Decimal(state.state) for state in states if state.state not in ignored_states
         )
         self._attr_native_value = round(summed, self._rounding_digits)
         self.async_schedule_update_ha_state(True)
